requirements.txt
Django>=5.0.6
OSlash>=0.6.3
annotated-types>=0.7.0
anyio>=4.4.0
asgiref>=3.8.1
attrs>=23.2.0
certifi>=2024.6.2
charset-normalizer>=3.3.2
django-modern-rpc>=1.0.3
django-ninja>=1.1.0
django-recaptcha>=4.0.0
h11>=0.14.0
httpcore>=1.0.5
httpx>=0.27.0
idna>=3.7
json-rpc>=1.15.0
jsonrpcserver>=5.0.9
jsonschema>=4.22.0
jsonschema-specifications>=2023.12.1
pillow>=10.3.0
pip>=24.0
psycopg2>=2.9.9
pydantic>=2.7.1
pydantic_core>=2.18.2
python-decouple>=3.8
python-telegram-bot>=21.3
referencing>=0.35.1
requests>=2.32.3
rpds-py>=0.18.1
sniffio>=1.3.1
sqlparse>=0.5.0
typing_extensions>=4.11.0
tzdata>=2024.1
urllib3>=2.2.1